refactor(app_partner): party ↔ ticket 양방향 커플링 해소
ticket feature가 party controller를 직접 import하던 구조를
ticket-local provider로 대체하고, TicketForm을 공유 위젯으로 이동.

- ticket_data_providers.dart 신규: ticketEntryGroupsProvider,
  ticketTemplatesByPartyProvider, ticketsByEventProvider
- TicketForm을 widgets/ 공유 영역으로 이동
- ticket → party, party → ticket 양방향 import 모두 제거

Closes #454

Constraint: TicketForm은 PartyEntryGroup 모델에 의존하나 minglit_kit에 정의되어 있어 허용
Rejected: TicketForm을 minglit_kit로 이동 | app-specific l10n 의존으로 불가
Confidence: high
Scope-risk: narrow
Not-tested: ticketEntryGroupsProvider의 partyRepository.getPartyById null 반환 케이스
